rename functions and coreData
