+++ ios/evy/Utils/Functions.swift
@@ -1,5 +1,5 @@
 //
-//  EVYFunctions.swift
+//  Functions.swift
 //  evy
 //
 //  Created by Geoffroy Lesage on 17/12/2023.
