+++ web/app/utils/functions.ts
